Move UInt*Range to chia.util.streamable
