+++ chia/wallet/wallet_coin_store.py
@@ -11,8 +11,7 @@ from chia.util.db_wrapper import DBWrapper2, execute_fetchone
 from chia.util.hash import std_hash
 from chia.util.ints import uint8, uint32, uint64
 from chia.util.lru_cache import LRUCache
-from chia.util.misc import UInt32Range, UInt64Range
-from chia.util.streamable import Streamable, VersionedBlob, streamable
+from chia.util.streamable import Streamable, UInt32Range, UInt64Range, VersionedBlob, streamable
 from chia.wallet.util.query_filter import AmountFilter, FilterMode, HashFilter
 from chia.wallet.util.wallet_types import CoinType, WalletType
 from chia.wallet.wallet_coin_record import WalletCoinRecord
